created monochrome and RGB backlight test
This will be less confusing for the techs. I will call the appropriate
function for the type of screen...

not sure what to do about the 16x4 though.: both the RGB and monochrome
version of that share a testbed, so they will probably have some
confusing messages pop up if the screen is monochrome, but I still send
commands for RGB.

Also, I'm noticing that the brightness settings are backwards. The auto
message text is aligned with the commands, however the actual brightness
of the backlight is reversed.