updating switch circuit and re-routing
the pi_grammer has the isolation switch, so I am now using the onboard switch to isolate the FJ's SPI lines. This will hopefully elliminate the need for the occasional hard reset on the testbed.

Re-routing everything manually. The original design was autorouted and was pretty ugly.

Currently most of the important stuff is routed nicely. Still needs LEDs and some switch stuff routed.